+++ examples/data-sources/exoscale_vpc/data-source.tf
@@ -1,0 +1,8 @@
+data "exoscale_vpc" "my_vpc" {
+  zone = "ch-gva-2"
+  name = "my-vpc"
+}
+
+output "my_vpc_id" {
+  value = data.exoscale_vpc.my_vpc.id
+}